feat: adding testnet faucet (#131)
* feat: adding benchmarks

* refac: restructuring lib.rs

* feat: implementing faucet

* fix: call index

* cargo fmt

* cargo fmt

---------
