+++ pallets/faucet/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pallet-governance"
+name = "pallet-faucet"
 version = "1.0.0"
 description = "FRAME pallet for runtime logic of Subspace Blockchain."
 authors = ["Commune Nucleus Team"]
@@ -8,9 +8,6 @@ edition = "2021"
 license = "Unlicense"
 publish = false
 repository = "https://github.com/commune-ai/subspace"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
 
 [lints.clippy]
 arithmetic_side_effects = "deny"
@@ -23,20 +20,17 @@ default = ["std"]
 std = [
     "frame-support/std",
     "frame-system/std",
-    "frame-benchmarking/std",
     "parity-scale-codec/std",
     "pallet-subspace/std",
     "scale-info/std",
     "sp-std/std",
 ]
 try-runtime = ["frame-support/try-runtime"]
-runtime-benchmarks = ["frame-benchmarking/runtime-benchmarks"]
 
 [dependencies]
 bty.workspace = true
 log.workspace = true
 
-pallet-governance-api = { path = "../governance/api", default-features = false }
 pallet-subspace = { path = "../subspace", default-features = false }
 
 frame-support.workspace = true
@@ -44,6 +38,7 @@ frame-system.workspace = true
 
 sp-runtime.workspace = true
 sp-std.workspace = true
+sp-core.workspace = true
 
 substrate-fixed.workspace = true
 
@@ -54,11 +49,7 @@ scale-info = { workspace = true, default-features = false, features = [
     "derive",
 ] }
 
-frame-benchmarking = { workspace = true, optional = true }
-
-
 [dev-dependencies]
 pallet-balances = { workspace = true, features = ["std"] }
-pallet-governance-api = { path = "./api", default-features = false }
 sp-io.workspace = true
 sp-tracing.workspace = true
